Refactored to match main
