+++ client/src/components/Listings/Listing/ListingModal.jsx
@@ -1,10 +1,10 @@
 import React, { useState } from 'react';
 import { useSelector, useDispatch } from 'react-redux';
-import { deleteListing } from '../../features/listings/listingsSlice';
-import ListingForm from '../Listings/ListingForm';
+import { deleteListing } from '../../../features/listings/listingsSlice';
+import ListingForm from '../ListingForm';
 import AliceCarousel from 'react-alice-carousel';
 import 'react-alice-carousel/lib/alice-carousel.css';
-import PoolForm from '../Pools/PoolForm';
+import PoolForm from '../../Pools/PoolForm';
 
 const ListingModal = ({ selectedListing, setSelectedListing }) => {
   const user = useSelector((store) => store.auth.token);
