+++ client/src/components/Map/Map.jsx
@@ -3,9 +3,9 @@ import { GoogleMap, useJsApiLoader, Marker } from '@react-google-maps/api';
 import Geocode from 'react-geocode';
 import Modal from 'react-modal';
 import { useSelector, useDispatch } from 'react-redux';
-import Listing from '../Listing/Listing';
+import Listing from '../Listings/Listing/Listing';
 import { fetchListings, getListingsData } from '../../features/listings/listingsSlice';
-import ListingModal from '../Listing/ListingModal';
+import ListingModal from '../Listings/Listing/ListingModal';
 
 const containerStyle = {
   width: '100vw',
